Started to implement descriptor allocator

+++ PathFinder/Source/HardwareAbstractionLayer/Descriptor.cpp
@@ -9,7 +9,7 @@ namespace HAL
 
     CPUDescriptor::~CPUDescriptor() {}
 
-    GPUDescriptor::GPUDescriptor(D3D12_CPU_DESCRIPTOR_HANDLE cpuHandle, D3D12_GPU_DESCRIPTOR_HANDLE gpuHandle, uint32_t indexInHeapRange)
+    GPUDescriptor::GPUDescriptor(D3D12_CPU_DESCRIPTOR_HANDLE cpuHandle, D3D12_GPU_DESCRIPTOR_HANDLE gpuHandle, uint64_t indexInHeapRange)
         : CPUDescriptor(cpuHandle), mGPUHandle{ gpuHandle }, mIndexInHeapRange{ indexInHeapRange } {}
 
     GPUDescriptor::~GPUDescriptor() {}
