+++ PathFinder/Source/HardwareAbstractionLayer/Descriptor.hpp
@@ -22,12 +22,12 @@ namespace HAL
 
     class GPUDescriptor : public CPUDescriptor {
     public:
-        GPUDescriptor(D3D12_CPU_DESCRIPTOR_HANDLE cpuHandle, D3D12_GPU_DESCRIPTOR_HANDLE gpuHandle, uint32_t indexInHeapRange);
+        GPUDescriptor(D3D12_CPU_DESCRIPTOR_HANDLE cpuHandle, D3D12_GPU_DESCRIPTOR_HANDLE gpuHandle, uint64_t indexInHeapRange);
         virtual ~GPUDescriptor() = 0;
 
     private:
         D3D12_GPU_DESCRIPTOR_HANDLE mGPUHandle;
-        uint32_t mIndexInHeapRange;
+        uint64_t mIndexInHeapRange;
 
     public:
         inline const auto& GPUHandle() const { return mGPUHandle; }
